my-flask-app/requirements.txt
Flask
flask-cors
tensorflow
Pillow
numpy